api_yamdb/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.16
django-import-export==2.7.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
django-filter==21.1
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
tablib==3.2.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
